(refactor): Move AppResourceTree test utils to separate file
