+++ specifyweb/frontend/js_src/lib/components/AppResources/__tests__/findAppResourceDirectory.test.ts
@@ -1,113 +1,15 @@
 import { requireContext } from "../../../tests/helpers";
 import { f } from "../../../utils/functools";
-import type { R, RA } from "../../../utils/types";
-import { filterArray, localized } from "../../../utils/types";
-import { serializeResource } from "../../DataModel/serializers";
-import { tables } from "../../DataModel/tables";
+import type { R } from "../../../utils/types";
+import { filterArray } from "../../../utils/types";
 import { findAppResourceDirectory, findAppResourceDirectoryKey } from "../Create";
 import type { AppResourcesTree } from "../hooks";
 import type { ScopedAppResourceDir } from "../types";
-
+import { utilsForTests } from "./utils";
+    
 requireContext();
 
-const makeAppResourceNode = (
-    label: string, 
-    key: string, 
-    directory: ScopedAppResourceDir | undefined, 
-    subCategories: AppResourcesTree
-): AppResourcesTree[number] => ({
-    label: localized(label),
-    key,
-    directory,
-    subCategories,
-    appResources: [],
-    viewSets: []
-});
-    
-const makeDirectory = (id: number): ScopedAppResourceDir => {
-    const dir = new tables.SpAppResourceDir.Resource({
-        id,
-        isPersonal: false,
-        collection: "/api/specify/collection/32768/",
-        discipline: "/api/specify/discipline/3/"
-    });
-
-    return ({...serializeResource(dir), scope: 'collection'})
-};
-
-// Make it part of functools?
-function *incrementor(){
-    let index = 0;
-    while (true){
-        yield index++;
-    }
-}
-
-type Incrementor = ReturnType<typeof incrementor>;
-
-function prefixIncrmentor(prefix: string, generator: ReturnType<typeof incrementor>){
-    return `${prefix}${generator.next().value}`
-}
-
-// This makes adding tests a bit easier.
-type Node = {
-    readonly id?: number;
-    readonly children: RA<Node>
-};
-
-const treeStructure: RA<Node> = [
-    {
-        id: 0, 
-        children: [
-            {
-                id: 0,
-                children: [
-                    {id: 0, children: []},
-                    {id: undefined, children: []}
-                ]
-            },
-            {
-                id: 0,
-                children: [
-                    {id: undefined, children: []},
-                    {id: undefined, children: []}
-                ]
-            }
-        ]
-    },
-    {
-        id: 0, 
-        children: [
-            {
-                id: 0,
-                children: [
-                    {id: 0, children: []},
-                    {id: undefined, children: []}
-                ]
-            },
-            {
-                id: 0,
-                children: [
-                    {id: 0, children: [
-                        {id: undefined, children: []}
-                    ]},
-                    {id: 0, children: []}
-                ]
-            }
-        ]
-    }
-];
-
-const makeTree = (nodes: RA<Node>, labelIncrementor: Incrementor, keyIncrementor: Incrementor, idIncrementor: Incrementor) : AppResourcesTree => (
-    nodes.map((node)=>makeAppResourceNode(
-        prefixIncrmentor("TestLabel", labelIncrementor) , 
-        prefixIncrmentor("TestKey", keyIncrementor) ,
-        node.id === undefined ? undefined : makeDirectory(idIncrementor.next().value as number),
-        makeTree(node.children, labelIncrementor, keyIncrementor, idIncrementor)
-    ))
-);
-
-
+const { treeStructure, makeAppResourceNode, makeDirectory, incrementor, makeTree } = utilsForTests;
 const simpleTree = () => [
             makeAppResourceNode("TestLabel", "TestKey1", makeDirectory(1), []),
             makeAppResourceNode("TestLabel2", "TestKey2", makeDirectory(2), []),
